fixed segault free

+++ exec/pipeline/pl_helper.c
@@ -6,7 +6,7 @@
 /*   By: eprottun <eprottun@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/02 17:08:27 by eprottun          #+#    #+#             */
-/*   Updated: 2025/10/03 12:49:33 by eprottun         ###   ########.fr       */
+/*   Updated: 2025/10/03 14:53:03 by eprottun         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -70,7 +70,7 @@ int cmd_tokens(t_cmd *current)
 {
 	t_entry *node;
 	size_t cmd_tokens;
-	size_t iter;
+	int iter;
 
 	node = current->line;
 	cmd_tokens = 0;
@@ -91,7 +91,7 @@ int cmd_init(t_cmd *current)
 {
 	t_entry *node;
 	size_t cmd_iter;
-	size_t iter;
+	int iter;
 
 	current->argv = malloc((cmd_tokens(current) + 1) * sizeof(char *));
 	if (!current->argv)
